Move album removal code to tagger
